+++ packages/core/dynamic-table/src/components/rankable/TableRow.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
 import { RankableTableBodyRow } from '../../styled/rankable/TableRow';
 import { HeadType, RowType } from '../../types';