+++ packages/core/inline-edit/src/components/InlineEditableTextfield.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Textfield from '@atlaskit/textfield';
 import ErrorIcon from '@atlaskit/icon/glyph/error';
 import { colors } from '@atlaskit/theme';