+++ packages/core/select/src/components/indicators.js
@@ -1,6 +1,6 @@
 // @flow
-
-import React from 'react';
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
 import { components } from 'react-select';
 import Spinner from '@atlaskit/spinner';
 import SelectClearIcon from '@atlaskit/icon/glyph/select-clear';
@@ -19,7 +19,7 @@ export const DropdownIndicator = (props: any) => (
 );
 
 export const LoadingIndicator = (props: any) => (
-  <div style={props.getStyles('loadingIndicator', props)} {...props.innerProps}>
+  <div css={props.getStyles('loadingIndicator', props)} {...props.innerProps}>
     <Spinner size="small" />
   </div>
 );