+++ packages/elements/mention/src/components/MentionPicker/index.tsx
@@ -302,7 +302,7 @@ export class MentionPicker extends React.PureComponent<
   }
 }
 
-export const MentionPickerWithAnalytics = withAnalyticsEvents({})(
+export const MentionPickerWithAnalytics = withAnalyticsEvents<Props>({})(
   MentionPicker,
 );