+++ packages/elements/mention/src/components/MentionSpotlight/index.tsx
@@ -1,16 +1,25 @@
-import * as React from 'react';
-import * as Styled from './styled';
+import React from 'react';
 
 import Button from '@atlaskit/button';
-import CloseIcon from '@atlaskit/icon/glyph/cross';
-
+import EditorCloseIcon from '@atlaskit/icon/glyph/editor/close';
 import Tooltip from '@atlaskit/tooltip';
+
+import {
+  SpotlightTitle,
+  SpotlightCloseTooltip,
+  SpotlightDescription,
+  SpotlightDescriptionLink,
+} from '../../util/i18n';
+import * as Styled from './styles';
 
 export interface Props {
   createTeamLink: string;
   /** Callback to track the event where user click on x icon */
   onClose: () => void;
 }
+
+const ICON_URL =
+  'https://ptc-directory-sited-static.us-east-1.prod.public.atl-paas.net/teams/avatars/2.svg';
 
 export default class MentionSpotlight extends React.Component<Props, {}> {
   render() {
@@ -20,33 +29,45 @@ export default class MentionSpotlight extends React.Component<Props, {}> {
       <Styled.Card>
         <Styled.Content>
           <Styled.Aside>
-            <img src={Styled.iconUrl} height={32} />
+            <img src={ICON_URL} height={32} />
           </Styled.Aside>
           <Styled.Section>
             <Styled.Heading>
               <Styled.Title>
-                <strong>You can now mention teams! </strong>
+                <SpotlightTitle />
               </Styled.Title>
             </Styled.Heading>
             <Styled.Body>
-              <p>
-                If you don't have any teams,
-                <a href={createTeamLink} target="_blank">
-                  {' '}
-                  start a team{' '}
-                </a>
-                with the group of people you work with daily.
-              </p>
+              <SpotlightDescription>
+                {description => (
+                  <p>
+                    {description}
+                    <SpotlightDescriptionLink>
+                      {linkText => (
+                        <a href={createTeamLink} target="_blank">
+                          {' '}
+                          {linkText}
+                        </a>
+                      )}
+                    </SpotlightDescriptionLink>
+                  </p>
+                )}
+              </SpotlightDescription>
             </Styled.Body>
           </Styled.Section>
           <Styled.Actions>
-            <Tooltip content="Close" position="bottom">
-              <Button
-                appearance="subtle"
-                iconBefore={<CloseIcon label="Close Modal" size="small" />}
-                onClick={onClose}
-              />
-            </Tooltip>
+            <SpotlightCloseTooltip>
+              {tooltip => (
+                <Tooltip content={tooltip} position="bottom">
+                  <Button
+                    appearance="subtle"
+                    iconBefore={<EditorCloseIcon label="Close" size="medium" />}
+                    onClick={onClose}
+                    spacing="none"
+                  />
+                </Tooltip>
+              )}
+            </SpotlightCloseTooltip>
           </Styled.Actions>
         </Styled.Content>
       </Styled.Card>