+++ packages/elements/share/src/i18n/pt_BR.ts
@@ -10,20 +10,35 @@ export default {
   'fabric.elements.share.form.send': 'Enviar',
   'fabric.elements.share.form.retry': 'Tentar novamente',
   'fabric.elements.share.form.comment.placeholder': 'Adicione uma mensagem',
-  'fabric.elements.share.form.user-picker.placeholder':
-    'Insira um nome, equipe ou e-mail',
+  'fabric.elements.share.form.user-picker.placeholder.generic':
+    'Insira um nome, uma equipe ou um e-mail',
+  'fabric.elements.share.form.user-picker.placeholder.existingUserOnly':
+    'Insira um nome ou uma equipe',
   'fabric.elements.share.form.user-picker.add-more': 'Insira mais itens',
   'fabric.elements.share.form.user-picker.validation.required':
     'Selecione pelo menos um usuário, equipe ou e-mail.',
-  'fabric.elements.share.form.user-picker.no-options.emptyQuery':
-    'Não encontramos nenhum resultado. Convide pessoas usando um endereço de e-mail.',
-  'fabric.elements.share.form.user-picker.no-options':
+  'fabric.elements.share.form.user-picker.no-options.domainBasedUserOnly':
+    'Você só pode convidar pessoas com endereços de e-mail {domains}. Entre em contato com o administrador do site para convidar outras pessoas.',
+  'fabric.elements.share.form.user-picker.no-options.existingUserOnly':
+    'Não encontramos nenhum resultado para "{inputValue}".',
+  'fabric.elements.share.form.user-picker.no-options.generic':
     'Não encontramos nenhum resultado para "{inputValue}". Convide pessoas usando um endereço de e-mail.',
+  'fabric.elements.share.trigger.button.icon.label':
+    'Ícone de compartilhamento',
   'fabric.elements.share.trigger.button.text': 'Compartilhar',
+  'fabric.elements.share.trigger.button.invite.text': 'Convidar',
+  'fabric.elements.share.copylink.button.icon.label': 'Ícone de cópia de link',
   'fabric.elements.share.copylink.button.text': 'Copiar link',
+  'fabric.elements.share.copied.to.clipboard.icon.label':
+    'Ícone de sucesso da cópia de link',
   'fabric.elements.share.copied.to.clipboard.message':
     'Link copiado para a área de transferência',
   'fabric.elements.share.form.capabilities.info.message':
     'Seu convite será enviado para a aprovação de um administrador',
+  'fabric.elements.share.failure.icon.label': 'Ícone de erro',
   'fabric.elements.share.failure.message': 'Não foi possível compartilhar',
+  'fabric.elements.share.success.message':
+    '{object, select, blogpost {Blog compartilhado} board {Painel compartilhado} calendar {Calendário compartilhado} draft {Rascunho compartilhado} filter {Filtro compartilhado} issue {Item compartilhado} media {Mídia compartilhada} page {Página compartilhada} project {Projeto compartilhado} pullrequest {Solicitação pull compartilhada} question {Pergunta compartilhada} report {Relatório compartilhado} repository {Repositório compartilhado} request {Solicitação compartilhada} roadmap {Roteiro compartilhado} site {Site compartilhado} space {Espaço compartilhado} other {Link compartilhado}}',
+  'fabric.elements.share.admin.notified.message':
+    'Seu administrador foi notificado',
 };