+++ packages/elements/share/src/i18n/ru.ts
@@ -10,20 +10,35 @@ export default {
   'fabric.elements.share.form.send': 'Отправить',
   'fabric.elements.share.form.retry': 'Еще раз',
   'fabric.elements.share.form.comment.placeholder': 'Введите сообщение',
-  'fabric.elements.share.form.user-picker.placeholder':
-    'Укажите имя, команду или адрес эл. почты',
+  'fabric.elements.share.form.user-picker.placeholder.generic':
+    'Укажите имя, название команды или адрес эл. почты',
+  'fabric.elements.share.form.user-picker.placeholder.existingUserOnly':
+    'Укажите имя или название команды',
   'fabric.elements.share.form.user-picker.add-more': 'Добавляйте еще',
   'fabric.elements.share.form.user-picker.validation.required':
     'Выберите хотя бы одного пользователя, команду или адрес эл. почты.',
-  'fabric.elements.share.form.user-picker.no-options.emptyQuery':
-    'Мы никого не нашли. Пригласите пользователей по адресу электронной почты.',
-  'fabric.elements.share.form.user-picker.no-options':
+  'fabric.elements.share.form.user-picker.no-options.domainBasedUserOnly':
+    'Вы можете приглашать только людей с адресами {domains}. Чтобы пригласить других пользователей, обратитесь к администратору сайта.',
+  'fabric.elements.share.form.user-picker.no-options.existingUserOnly':
+    'Мы никого не нашли по запросу «{inputValue}».',
+  'fabric.elements.share.form.user-picker.no-options.generic':
     'Мы никого не нашли по запросу «{inputValue}». Пригласите пользователей по адресу электронной почты.',
+  'fabric.elements.share.trigger.button.icon.label': 'Значок «Поделиться»',
   'fabric.elements.share.trigger.button.text': 'Поделиться',
+  'fabric.elements.share.trigger.button.invite.text': 'Пригласить',
+  'fabric.elements.share.copylink.button.icon.label':
+    'Значок «Копировать ссылку»',
   'fabric.elements.share.copylink.button.text': 'Копировать ссылку',
+  'fabric.elements.share.copied.to.clipboard.icon.label':
+    'Значок скопированной ссылки',
   'fabric.elements.share.copied.to.clipboard.message':
     'Ссылка скопирована в буфер обмена',
   'fabric.elements.share.form.capabilities.info.message':
     'Приглашение будет отправлено на рассмотрение администратору',
+  'fabric.elements.share.failure.icon.label': 'Значок ошибки',
   'fabric.elements.share.failure.message': 'Не удалось поделиться',
+  'fabric.elements.share.success.message':
+    '{object, select, blogpost {Blog post shared} {Board shared} к доске {Calendar shared} к календарю {Draft shared} к черновику {Filter shared} к фильтру {Issue shared} к задаче {Media shared} к мультимедийному файлу {Page shared} к странице {Project shared} к проекту {Pull Request shared} к запросу pull {Question shared} к вопросу {Report shared} к отчету {Repository shared} к репозиторию {Request shared} к запросу {Roadmap shared} к дорожной карте {Site shared} к сайту {Space shared} к разделу {Link Shared} к ссылке}',
+  'fabric.elements.share.admin.notified.message':
+    'Ваш администратор получил уведомление',
 };