+++ packages/monorepo-tooling/dependency-version-analytics/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  populateProduct,
+  populatePackage,
+} from './commands/populate-historic-data';
+export { sendAnalytics } from './util/analytics';
+export * from './types';