+++ packages/core/analytics-next/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @atlaskit/analytics-next
 
+## 5.1.0
+
+### Minor Changes
+
+- [minor][d0db01b410](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d0db01b410):
+
+  TypeScript users of withAnalyticsEvents and withAnalyticsContext are now required to provide props as a generic type. This is so that TypeScript can correctly calculate the props and defaultProps of the returned component.
+
+  Before:
+
+  ```typescript
+  withAnalyticsEvents()(Button) as ComponentClass<Props>;
+  ```
+
+  After:
+
+  ```typescript
+  withAnalyticsEvents<Props>()(Button);
+  ```
+
 ## 5.0.2
 
 ### Patch Changes