+++ packages/core/select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/select
 
+## 10.0.0
+
+### Major Changes
+
+- [major][790e66bece](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/790e66bece):
+
+  Major bump to react-select which includes a bump from emotion 9 --> 10, this will impact users who are currently creating custom components using emotion. Empty values in selects have also now been changed to be more deterministic across single and multi select. See https://github.com/JedWatson/react-select/issues/3585 for details
+
 ## 9.1.10
 
 - Updated dependencies [87a2638655](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/87a2638655):