+++ packages/editor/email-renderer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/email-renderer
 
+## 2.8.0
+
+### Minor Changes
+
+- [minor][4c3772ce61](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4c3772ce61):
+
+  CS-1238 Added generic icon for media attachments
+
+## 2.7.1
+
+### Patch Changes
+
+- [patch][59fb844cd5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/59fb844cd5):
+
+  CS-1184 Email renderer - prevent tables from flowing outside container
+
 ## 2.7.0
 
 ### Minor Changes