+++ packages/search/quick-search/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/quick-search
 
+## 7.6.2
+
+### Patch Changes
+
+- [patch][5427f7028a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5427f7028a):
+
+  Fixing problem with subtext colours not matching between pages and people in the Quick Search complex experiment.
+
+## 7.6.1
+
+### Patch Changes
+
+- [patch][8f711664af](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8f711664af):
+
+  Added analytics for current space filter component
+
 ## 7.6.0
 
 ### Minor Changes